✨ (grabber): add MangamonkChromedp implementation for scraping mangamonk.com Introduce a new grabber for mangamonk.com that utilizes browserless/chromedp for scraping. This implementation includes methods for testing URLs, fetching titles, chapters, and images, enhancing the overall functionality of the comic downloader.
🗑️ (grabber): remove plainhtml.go file

The plainhtml.go file is deleted as it is no longer needed in the project. This helps to clean up the codebase and remove unused components.

✨ (grabber): add support for reaperscans.com API
Implement a new grabber for reaperscans.com to download comics using its API. This addition enhances the functionality of the comic downloader by allowing users to access content from reaperscans.com.

🗑️ (grabber): remove tcb.go file

The removal of the tcb.go file is done to clean up unused code and improve project maintainability.